Renamed utils to common
